InLineNotification with severity property
