+++ src/components/InLineNotification.tsx
@@ -1,9 +1,27 @@
 import React from 'react'
-import { StyleSheet, Text, TextStyle, View, ViewStyle } from 'react-native'
+import { StyleProp, StyleSheet, Text, TextStyle, View, ViewStyle } from 'react-native'
 import AttentionIcon from 'src/icons/Attention'
 import Colors from 'src/styles/colors'
 import fontStyles, { typeScale } from 'src/styles/fonts'
 import { Spacing } from 'src/styles/styles'
+
+export enum Severity {
+  Informational,
+  Warning,
+  Error,
+}
+
+interface Props {
+  severity: Severity
+  title?: string
+  description: string
+  style?: StyleProp<ViewStyle>
+  ctaLabel?: string
+  onPressCta?: () => void
+  ctaLabel2?: string
+  onPressCta2?: () => void
+  testID?: string
+}
 
 interface CustomStyles {
   container: ViewStyle
@@ -12,35 +30,24 @@ interface CustomStyles {
   ctaLabel2: TextStyle
 }
 
-interface Props {
-  title?: string
-  description: string
-  styles: Partial<CustomStyles>
-  ctaLabel?: string
-  onPressCta?: () => void
-  ctaLabel2?: string
-  onPressCta2?: () => void
-  testID?: string
-}
-
-type PropsWithPresetStyle = Omit<Props, 'styles'>
-
-export function Notification({
+export function InLineNotification({
+  severity,
   title,
   description,
-  styles,
+  style,
   ctaLabel,
   onPressCta,
   ctaLabel2,
   onPressCta2,
   testID,
 }: Props) {
+  const severityStyle = severityStyles[severity]
   return (
-    <View style={[defaultStyles.container, styles?.container]} testID={testID}>
+    <View style={[defaultStyles.container, style]} testID={testID}>
       {title && (
         <View style={defaultStyles.row}>
           <View style={defaultStyles.attentionIcon}>
-            <AttentionIcon color={styles?.attentionIcon?.color ?? Colors.informationDark} />
+            <AttentionIcon color={severityStyle.attentionIcon.color} />
           </View>
           <Text style={defaultStyles.titleText}>{title}</Text>
         </View>
@@ -49,9 +56,7 @@ export function Notification({
       <View style={defaultStyles.bodyContainer}>
         <View style={[defaultStyles.row, defaultStyles.bodyRow]}>
           <View style={defaultStyles.attentionIcon}>
-            {!title && (
-              <AttentionIcon color={styles?.attentionIcon?.color ?? Colors.informationDark} />
-            )}
+            {!title && <AttentionIcon color={severityStyle.attentionIcon.color} />}
           </View>
           <Text style={defaultStyles.bodyText}>{description}</Text>
         </View>
@@ -59,12 +64,15 @@ export function Notification({
         {(ctaLabel || ctaLabel2) && (
           <View style={[defaultStyles.row, defaultStyles.ctaRow]}>
             {ctaLabel && onPressCta && (
-              <Text style={[defaultStyles.ctaLabel, styles?.ctaLabel]} onPress={onPressCta}>
+              <Text style={[defaultStyles.ctaLabel, severityStyle.ctaLabel]} onPress={onPressCta}>
                 {ctaLabel}
               </Text>
             )}
             {ctaLabel2 && onPressCta2 && (
-              <Text style={[defaultStyles.ctaLabel2, styles?.ctaLabel2]} onPress={onPressCta2}>
+              <Text
+                style={[defaultStyles.ctaLabel2, severityStyle.ctaLabel2]}
+                onPress={onPressCta2}
+              >
                 {ctaLabel2}
               </Text>
             )}
@@ -126,31 +134,25 @@ const defaultStyles = StyleSheet.create({
   },
 })
 
-const informationalStyle: Partial<CustomStyles> = {
-  container: { backgroundColor: Colors.informationFaint },
-  attentionIcon: { color: Colors.informationDark },
-  ctaLabel: { color: Colors.informationDark },
-  ctaLabel2: { color: Colors.informationDark },
+const severityStyles: Record<Severity, CustomStyles> = {
+  [Severity.Informational]: {
+    container: { backgroundColor: Colors.informationFaint },
+    attentionIcon: { color: Colors.informationDark },
+    ctaLabel: { color: Colors.informationDark },
+    ctaLabel2: { color: Colors.informationDark },
+  },
+  [Severity.Warning]: {
+    container: { backgroundColor: Colors.warningLight },
+    attentionIcon: { color: Colors.warningDark },
+    ctaLabel: { color: Colors.warningDark },
+    ctaLabel2: { color: Colors.warningDark },
+  },
+  [Severity.Error]: {
+    container: { backgroundColor: Colors.errorLight },
+    attentionIcon: { color: Colors.errorDark },
+    ctaLabel: { color: Colors.errorDark },
+    ctaLabel2: { color: Colors.errorDark },
+  },
 }
 
-const warningStyle: Partial<CustomStyles> = {
-  container: { backgroundColor: Colors.warningLight },
-  attentionIcon: { color: Colors.warningDark },
-  ctaLabel: { color: Colors.warningDark },
-  ctaLabel2: { color: Colors.warningDark },
-}
-
-const errorStyle: Partial<CustomStyles> = {
-  container: { backgroundColor: Colors.errorLight },
-  attentionIcon: { color: Colors.errorDark },
-  ctaLabel: { color: Colors.errorDark },
-  ctaLabel2: { color: Colors.errorDark },
-}
-
-Notification.Informational = (props: PropsWithPresetStyle) =>
-  Notification({ ...props, styles: informationalStyle })
-Notification.Warning = (props: PropsWithPresetStyle) =>
-  Notification({ ...props, styles: warningStyle })
-Notification.Error = (props: PropsWithPresetStyle) => Notification({ ...props, styles: errorStyle })
-
-export default Notification
+export default InLineNotification
